reduxの基礎導入
