+++ src/App.tsx
@@ -1,11 +1,13 @@
 import React, { FC } from 'react';
 import { Register } from './component/register';
+import { Counter } from './component/counter';
 import { Layout } from './component/layout';
 
 const App: FC = () => {
   return (
     <Layout>
       <Register />
+      <Counter />
     </Layout>
   );
 };
