+++ src/interface.ts
@@ -2,4 +2,11 @@ export interface Store {
   counter: {
     count: number;
   };
+  forecast: {
+    date: Date;
+    summary: string;
+    temperatureMax: number;
+    temperatureMin: number;
+    icon: 'sun' | 'sun' | 'rain' | 'snowflake' | 'cloud' | 'question';
+  };
 }
